requirements.txt
litestar>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
websockets>=12.0.0
uvicorn>=0.27.1
pydantic>=2.6.1 